% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood-profiling.R
\name{binary_likelihood_profs}
\alias{binary_likelihood_profs}
\title{Return likelihood profiles for binary covariates special case}
\usage{
binary_likelihood_profs(
  best_pars,
  max_loglike,
  mc_samples_summary,
  alpha = 0.05,
  multiple_outside_transmissions = FALSE
)
}
\arguments{
\item{best_pars}{vector of MLE}

\item{max_loglike}{loglike value from MLE}

\item{mc_samples_summary}{tree summaries, output from \code{sample_mc_binary_cov}}

\item{alpha}{\eqn{alpha}-value to make a (1-\eqn{alpha})\% CI}

\item{multiple_outside_transmissions}{logical indicating to use base model
or multiple outside transmissions model.}
}
\value{
matrix of dimension (length of best_pars) x 3 where the columns are the best parameter estimate,
the lower CI (\code{lower}) and upper CI (\code{upper})
}
\description{
Return likelihood profiles for binary covariates special case
}
