% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood-profiling.R
\name{binary_loglike_wrapper}
\alias{binary_loglike_wrapper}
\title{Wrapper to optimize over a single variable for likelihood profiling}
\usage{
binary_loglike_wrapper(
  x,
  max_loglike,
  best_pars,
  obs_data_summary,
  mc_samples_summary,
  alpha = 0.05,
  beta_index = 1,
  multiple_outside_transmissions = FALSE
)
}
\arguments{
\item{x}{parameter to optimize over}

\item{max_loglike}{maximum loglike value from output}

\item{best_pars}{vector of best parameters}

\item{obs_data_summary}{actual data summary with cluster_size, x_pos, and x_neg and frequency of each}

\item{mc_samples_summary}{output of \code{sample_mc_binary_cov}}

\item{alpha}{\eqn{\alpha}-value for computing confidence intervals.}

\item{beta_index}{which parameter to optimize over.}

\item{multiple_outside_transmissions}{whether to use base model or not}
}
\value{
zeros of likelihood profiling for given \eqn{alpha}
}
\description{
Wrapper to optimize over a single variable for likelihood profiling
}
\details{
internal function only
}
