% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-clusters.R
\name{bootstrap_clusters}
\alias{bootstrap_clusters}
\title{Resample clusters from an observed data set}
\usage{
bootstrap_clusters(clusters, n_boot = NULL)
}
\arguments{
\item{clusters}{data frame of observed data with the following columns
\describe{
\item{cluster_id}{cluster ID}
\item{other covariates}{any other covariates}
}}

\item{n_boot}{number of clusters to resample.  If NULL, will re-sample as many in original cluster}
}
\value{
data frame with new cluster_id variable but can be traced back to original cluster
through the column \code{original_id}
}
\description{
Resample clusters from an observed data set
}
\examples{
  data <- data.frame(cluster_id = c("a", "a", "a",
"b", "b",
"c", "c",
"d"),
x = c(1, 0, 1,
     1, 1,
     0, 1,
     0))
boot_clusters <- bootstrap_clusters(clusters = data)
boot_clusters
}
