% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-A-infects-B.R
\name{calculate_transmission_prob}
\alias{calculate_transmission_prob}
\title{Get the probability of transmission given a set of covariates}
\usage{
calculate_transmission_prob(
  data,
  inf_params,
  covariate_names = names(inf_params)[-1]
)
}
\arguments{
\item{data}{data frame with variables corresponding to those in cov_names}

\item{inf_params}{values of parameters, possibly named}

\item{covariate_names}{covariate names}
}
\value{
probability of infection, the result of a logit function
}
\description{
Get the probability of transmission given a set of covariates
}
