% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-A-infects-B.R
\name{compute_A_infects_B}
\alias{compute_A_infects_B}
\title{Compute probability from samples that person A infects person B}
\usage{
compute_A_infects_B(df)
}
\arguments{
\item{df}{data frame with the following columns
\describe{
\item{cluster_id}{unique ID of cluster}
\item{prob_trans}{individual probability of transmission}
\item{like}{likelihood of cluster}
\item{id}{id of person within cluster}
\item{inf_id}{infector ID of person}
\item{feature_id}{a unique person/or set of features across the MC clusters}
}}
}
\value{
a n x n matrix where n is the number of individuals in the cluster.
The rows correspond to the INFECTOR id and the columns to the INFECTED id.
So entry i,j is the probability i infects j
}
\description{
Compute probability from samples that person A infects person B
}
