% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-loglike.R
\name{covariate_df_to_mat}
\alias{covariate_df_to_mat}
\title{Take data frame and turn it into matrix for logistic regression}
\usage{
covariate_df_to_mat(sampled_data, cov_names)
}
\arguments{
\item{sampled_data}{data frame with cov_names}

\item{cov_names}{null or string of vector}
}
\value{
matrix of dimension n x (p+1)
}
\description{
Take data frame and turn it into matrix for logistic regression
}
