% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_general_outbreaks.R
\name{draw_covariate_rows}
\alias{draw_covariate_rows}
\title{Randomly sample covariate rows}
\usage{
draw_covariate_rows(n, df, covariate_weights)
}
\arguments{
\item{n}{number rows to draw}

\item{df}{data frame to draw from}

\item{covariate_weights}{default is NULL which draws uniformly at random with replacement from the sample_covariates_df.  Otherwise, the weights are used.}
}
\value{
new df
}
\description{
Randomly sample covariate rows
}
