% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-loglike.R
\name{general_cluster_like.dt}
\alias{general_cluster_like.dt}
\title{Get the likelihood for a single cluster}
\usage{
general_cluster_like.dt(prob_inf, n_inf)
}
\arguments{
\item{prob_inf}{probability of onward transmission}

\item{n_inf}{number infected by this person}
}
\value{
a single number, the likelihood
}
\description{
Get the likelihood for a single cluster
}
\details{
(specifically for data.table)
}
