% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-loglike.R
\name{general_cluster_like}
\alias{general_cluster_like}
\title{Get the likelihood for a single cluster}
\usage{
general_cluster_like(data)
}
\arguments{
\item{data}{data frame with the following columns
\describe{
\item{prob_inf}{probability of onward transmission}
\item{n_inf}{number infected by this person}
}}
}
\value{
a single number, the likelihood
}
\description{
Get the likelihood for a single cluster
}
