% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_general_outbreaks.R
\name{general_generation_infection}
\alias{general_generation_infection}
\title{Infect the current generation given the previous generation}
\usage{
general_generation_infection(
  cluster_id,
  df,
  covariate_weights = NULL,
  gen,
  prev_gen,
  cluster_size,
  max_size
)
}
\arguments{
\item{cluster_id}{unique cluster ID}

\item{df}{covariate data frame.  One is prob_inf}

\item{covariate_weights}{either NULL or weight of covariates}

\item{gen}{new generation number}

\item{prev_gen}{data frame with at least smear status (0/1) and person_id}

\item{cluster_size}{size of previous cluster}

\item{max_size}{maximum size of cluster}
}
\value{
list with the following entries
\describe{
\item{new_cluster_size}{new cluster size with the current generation}
\item{cur_gen}{data frame of current generation (new infections)}
\item{n_inf}{number of infections for each person in the previous generation}
}
}
\description{
Infect the current generation given the previous generation
}
