% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-loglike.R
\name{general_loglike_mc_trees}
\alias{general_loglike_mc_trees}
\title{Calculate the log likelihood for a set of data, averaged by n}
\usage{
general_loglike_mc_trees(mc_trees)
}
\arguments{
\item{mc_trees}{data frame with either (the cluster_id, person_id, n_inf, cluster_size, prob_inf) or additionally with the covariate information}
}
\value{
data frame of average loglikelihood for a cluster of a given size
}
\description{
Calculate the log likelihood for a set of data, averaged by n
}
