% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-loglike.R
\name{general_loglike_summarized}
\alias{general_loglike_summarized}
\title{Calculate the estimated general loglikelihood}
\usage{
general_loglike_summarized(
  inf_params,
  obs_summarized_data,
  sampled_data,
  return_neg = TRUE,
  cov_mat = NULL,
  cov_names = NULL
)
}
\arguments{
\item{inf_params}{vector of p parameters}

\item{obs_summarized_data}{data frame with the following columns
\describe{
\item{cluster_size}{size of the cluster}
\item{freq}{frequency that cluster size occurred in the data}
}}

\item{sampled_data}{data frame of samples that correspond to the data.  See details.  This is the output of \code{sample_general_cond_trees()}.}

\item{return_neg}{default is TRUE.  Returns the negative loglike}

\item{cov_mat}{optional matrix of covariates corresponding to the sampled_data}

\item{cov_names}{covariate vector of length p which correspond in order to the betas}
}
\value{
estimated average loglikelihood for the observed data
}
\description{
Calculate the estimated general loglikelihood
}
\details{
This is a specialized log likelihood function where we first estimate the average log likelihood of trees conditioned by their total size through sampling.  This is very much dependent on the values of sampled_data.
}
