% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_general_cond_trees.R
\name{general_tree_sampler}
\alias{general_tree_sampler}
\title{Generate trees of certain size}
\usage{
general_tree_sampler(n, B)
}
\arguments{
\item{n}{size of cluster}

\item{B}{number of times to repeat}
}
\value{
data frame of trees
}
\description{
Generate trees of certain size
}
