% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood-profiling.R
\name{loglike_wrapper}
\alias{loglike_wrapper}
\title{Wrapper to optimize over a single variable for likelihood profiling}
\usage{
loglike_wrapper(
  x,
  max_loglike,
  best_pars,
  mc_trees,
  covariate_names,
  alpha = 0.05,
  beta_index = 1,
  multiple_outside_transmissions = FALSE
)
}
\arguments{
\item{x}{parameter to optimize over}

\item{max_loglike}{maximum loglike value from output}

\item{best_pars}{vector of best parameters}

\item{mc_trees}{output of \code{sample_mc_trees}}

\item{covariate_names}{covariate names to look over}

\item{alpha}{\eqn{\alpha}-value for computing confidence intervals.}

\item{beta_index}{which parameter to optimize over.}

\item{multiple_outside_transmissions}{whether to use base model or not}
}
\value{
zeros of likelihood profiling for given \eqn{alpha}
}
\description{
Wrapper to optimize over a single variable for likelihood profiling
}
\details{
internal function only.  This is for the general case
}
