% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-trees-fast.R
\name{sample_connections}
\alias{sample_connections}
\title{Sample connections and form a tree}
\usage{
sample_connections(gen_sizes)
}
\arguments{
\item{gen_sizes}{vector of generation sizes}
}
\value{
data frame with the following columns
\describe{
\item{gen}{generation number}
\item{n_in_gen}{number inside that generation}
\item{inf_id}{infector ID}
\item{id}{unique ID within cluster}
}
}
\description{
Sample connections and form a tree
}
