% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-general-cond-trees-outsider.R
\name{sample_gen_out_cond}
\alias{sample_gen_out_cond}
\title{Sample trees with same features as original but permuted through the trees}
\usage{
sample_gen_out_cond(observed_data, B)
}
\arguments{
\item{observed_data}{data.table frame with the following columns
\describe{
\item{cluster_id}{unique cluster ID}
\item{person_id}{unique identifier for every person}
\item{covariates}{covariates of the individuals}
}}

\item{B}{number of samples for each individual tree}
}
\value{
data frame with the following columns
\describe{
\item{cluster_id}{unique cluster_id for the sampled cluster}
\item{orig_id}{original ID of the given cluster}
\item{variable}{columns from featureless tree sampling}
\item{person_id}{original identifier}
}
}
\description{
Sample trees with same features as original but permuted through the trees
}
\details{
assumes an observed outside infection
}
