% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_general_cond_trees.R
\name{sample_general_cond_trees}
\alias{sample_general_cond_trees}
\title{Sample general conditional trees}
\usage{
sample_general_cond_trees(
  n_vec,
  B = 100,
  feature_type = "single_gauss",
  params_list = list(mean = 0, sd = 0.5),
  multiple_outside_transmissions = FALSE
)
}
\arguments{
\item{n_vec}{vector of size of clusters}

\item{B}{number of sampled trees for each size}

\item{feature_type}{method of how we sample covariates.  Currently only "single_gauss", "empirical" are supported}

\item{params_list}{of additional parameters}

\item{multiple_outside_transmissions}{logical indicating whether
we allow multiple outside transmissions or not.  Default is FALSE}
}
\value{
data frame of trees
}
\description{
Sample general conditional trees
}
