% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_general_cond_trees.R
\name{sample_general_tree_perm}
\alias{sample_general_tree_perm}
\title{Sample the actual tree for given permutation}
\usage{
sample_general_tree_perm(gen_sizes)
}
\arguments{
\item{gen_sizes}{vector of generation sizes}
}
\value{
sampled tree data framegiven the generation sizes
\describe{
\item{cluster_size}{cluster_size}
\item{id}{person ID}
\item{gen}{generation number}
\item{n_inf}{number infected}
}
}
\description{
Sample the actual tree for given permutation
}
