% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-mc-binary-cov.R
\name{sample_mc_binary_cov}
\alias{sample_mc_binary_cov}
\title{Sample binary covariate transmission trees}
\usage{
sample_mc_binary_cov(
  B,
  observed_cluster_summaries,
  multiple_outside_transmissions = FALSE
)
}
\arguments{
\item{B}{number of MC samples}

\item{observed_cluster_summaries}{data frame with the following columns
\describe{
\item{freq}{frequency of the following cluster}
\item{cluster_size}{size of the cluster}
\item{x_pos}{number of positive x (or 1)}
\item{x_neg}{number of negative x (or 0)}
}}

\item{multiple_outside_transmissions}{logical indicating whether to
sample from the multiple outside transmissions model.  Default is FALSE.}
}
\value{
a data frame of sampled transmission trees
with the following columns (the last one is only needed for the multiple
outsider model)
\describe{
\item{freq}{frequency of cluster with the following characteristics}
\item{cluster_size}{size of the cluster}
\item{x_pos}{number of x positives in the cluster}
\item{x_neg}{number of x negatives in the cluster}
\item{mc_freq}{frequency of MC trees with the following}
\item{x_pos_trans}{number of transmissions from a positive individual in the tree}
\item{x_neg_trans}{number of transmissions from a negative individual in the tree}
\item{root_node_sign}{whether root node is pos or neg}
}
@export
}
\description{
Sample binary covariate transmission trees
}
