% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-mc-binary-cov.R
\name{sample_mc_binary_cov_inner}
\alias{sample_mc_binary_cov_inner}
\title{Innfer function to sample MC trees with binary covariates}
\usage{
sample_mc_binary_cov_inner(x_pos, x_neg, B, root_node = NULL)
}
\arguments{
\item{x_pos}{number of positive individuals in cluster}

\item{x_neg}{number of negative individuals in cluster}

\item{B}{number of trees to sample}

\item{root_node}{if NULL, use the base model.  If 1, ensure that the root node is positive.  If 0, ensure that the root node is negative.}
}
\value{
a data frame with the following columns
\describe{
\item{freq}{frequency of the observed data}
\item{cluster_size}{total size of the cluster}
\item{x_pos}{number of individuals in the cluster with the feature of interest =1}
\item{x_neg}{number of individuals in the cluster with the feature of interest  =0}
\item{mc_freq}{number of MC frequency of clusters}
\item{x_pos_trans}{number of transmissions from a positive}
\item{x_neg_trans}{number of transmissions from a negative}
\item{root_node}{only for the multiple outsider.  Is 1 if the outsider is positive and 0 if it is negative}
}
}
\description{
Innfer function to sample MC trees with binary covariates
}
