% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_general_cond_trees.R
\name{sample_mc_trees}
\alias{sample_mc_trees}
\title{Sample trees with same features as original but permuted through the trees}
\usage{
sample_mc_trees(
  observed_data,
  B = 100,
  covariate_names = "x",
  multiple_outside_transmissions = FALSE,
  replace_vars = FALSE
)
}
\arguments{
\item{observed_data}{data frame with the following columns
\describe{
\item{cluster_id}{unique cluster ID}
\item{covariates}{covariates of the individuals}
}}

\item{B}{number of trees to sample}

\item{covariate_names}{string names of covariates}

\item{multiple_outside_transmissions}{logical indicating whether
we allow multiple outside transmissions or not.  Default is FALSE}

\item{replace_vars}{logical indicating whether we
can re sample from the covariate df.  Default is FALSE}
}
\value{
data frame with the following columns
}
\description{
Sample trees with same features as original but permuted through the trees
}
