% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-general-cond-trees-outsider.R
\name{sample_single_gen_out_cond}
\alias{sample_single_gen_out_cond}
\title{Make the sampled clusters for a single cluster}
\usage{
sample_single_gen_out_cond(cluster_person_ids, B, person_ids, orig_id)
}
\arguments{
\item{cluster_person_ids}{ids of the people WITHIN the cluster}

\item{B}{number of times to sample a cluster}

\item{person_ids}{person_ids of all people}

\item{orig_id}{original ID}
}
\value{
data frame of transmissions trees for the original cluster
}
\description{
Make the sampled clusters for a single cluster
}
