% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_general_cond_trees.R
\name{sample_trees_fixed_g}
\alias{sample_trees_fixed_g}
\title{Sample trees with given generation sizes}
\usage{
sample_trees_fixed_g(perm_mat)
}
\arguments{
\item{perm_mat}{g x B matrix where each column is a unique permutation of generation sizes}
}
\value{
data frame with the following columns
\describe{
\item{cluster_id}{unique cluster ID}
\item{id}{person ID}
\item{gen}{generation number}
\item{n_inf}{number infected by this person}
}
}
\description{
Sample trees with given generation sizes
}
