% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_general_cond_trees.R
\name{sample_unif_trees_no_features}
\alias{sample_unif_trees_no_features}
\title{Sample trees}
\usage{
sample_unif_trees_no_features(n_vec, B)
}
\arguments{
\item{n_vec}{vector of cluster sizes}

\item{B}{number of samples}
}
\value{
data frame of sampled trees (could be very large)
}
\description{
Sample trees
}
