% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-trees-fast.R
\name{sample_unique_perms}
\alias{sample_unique_perms}
\title{Sample unique permutations from a partition space}
\usage{
sample_unique_perms(g, n, B)
}
\arguments{
\item{g}{number of generations}

\item{n}{total size of cluster}

\item{B}{number of permutations to draw}
}
\value{
matrix of size g x B and the first row is all 1.  Each column is a unique permutation that sums to n and all are positive entries
}
\description{
Sample unique permutations from a partition space
}
