% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_general_outbreaks.R
\name{simulate_bp}
\alias{simulate_bp}
\title{Simulate the branching process of flipping until failure for K clusters}
\usage{
simulate_bp(
  K,
  inf_params,
  sample_covariates_df,
  covariate_names,
  covariate_weights = NULL,
  max_size = 50
)
}
\arguments{
\item{K}{number of total clusters to simulate}

\item{inf_params}{vector with beta coefficients to use in logistic function for probability of transmission}

\item{sample_covariates_df}{Data frame of covariates to sample from}

\item{covariate_names}{names of the covariates.  Must match size of inf_params - 1.}

\item{covariate_weights}{default is NULL which draws uniformly at random with replacement from the sample_covariates_df.  Otherwise, the weights are used.}

\item{max_size}{maximum size a cluster can be}
}
\value{
data frame with the following columns
\describe{
\item{cluster_id}{unique cluster ID}
\item{person_id}{order of infection in the cluster}
\item{gen}{generation number (>=0)}
\item{inf_id}{ID of the infector}
\item{n_inf}{number of people infected by person}
\item{censored}{whether the cluster end was censored or not}
\item{cluster_size}{size of the cluster}
\item{covariates}{covariates of the individuals}
}
}
\description{
Simulate the branching process of flipping until failure for K clusters
}
\details{
Generate a branching process according to the following process.
First a root infector is drawn covariates \eqn{X}
from some distribution $F$ (given by the set of covariates in \code{sample_covariates_df}) and has probability of
transmission according to a logit function.
The number of infections produced by the root node $N_{(1,1)}
is a geometric random variable with probability $p_{(1,1)}$
where the indexing represents $(g=$, generation, $i=$ index).
If $N_{(1,1)} > 0$, then the $N_{(1,1)}$ infections are added to
the cluster and assigned to generation $g=2$ with indices $i=1,
\dots, N_{(1,1)}$ and covariats are drawn for these new infections.
The infection process continues with individuals $(2, 1)$ through $(2, $N_{(1,1)})$
where new infections are added, in order to the subsequent generation.
The process terminates when either there are no new infections or the
maximum number of infections specified in \code{max_size} is reached.
\deqn{X_{(g,i)} \sim F}
\deqn{p_{(g,i)} = logit^{-1}\left ( X_{(g,i)} \beta\right )}
\deqn{N_{(g,i)} \sim Geometric(p_{(g,i)})}
}
\examples{
set.seed(2020)
inf_params <- c("beta_0" = -2, "beta_1" = 2)
df <- data.frame(x= c(0, 1))
branching_processes <- simulate_bp(K = 10,
inf_params = inf_params,
covariate_names = "x",
sample_covariates_df = df)
head(branching_processes)
table(branching_processes$cluster_size) /
sort(unique(branching_processes$cluster_size))
}
