% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_general_outbreaks.R
\name{simulate_general_outbreak}
\alias{simulate_general_outbreak}
\title{Simulate the process of flipping until failure for K clusters}
\usage{
simulate_general_outbreak(
  K,
  inf_params,
  sample_covariates_df,
  covariate_names,
  covariate_weights = NULL,
  max_size = 50
)
}
\arguments{
\item{K}{number of total clusters to simulate}

\item{inf_params}{vector with beta coefficients to use in logistic function for probability of transmission}

\item{sample_covariates_df}{Data frame of covariates to sample from}

\item{covariate_names}{names of the covariates.  Must match size of inf_params - 1.}

\item{covariate_weights}{default is NULL which draws uniformly at random with replacement from the sample_covariates_df.  Otherwise, the weights are used.}

\item{max_size}{maximum size a cluster can be}
}
\value{
data frame with the following columns
\describe{
\item{cluster_id}{unique cluster ID}
\item{person_id}{order of infection in the cluster}
\item{gen}{generation number (>=0)}
\item{inf_id}{ID of the infector}
\item{n_inf}{number of people infected by person}
\item{censored}{whether the cluster end was censored or not}
\item{cluster_size}{size of the cluster}
\item{covariates}{covariates of the individuals}
}
}
\description{
Simulate the process of flipping until failure for K clusters
}
\details{
breadth not depth.  Generate generation by generation as opposed to going up the branch til termination.
}
\examples{
inf_params <- c("beta_0" = -2, "beta_1" = 1)
}
