% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_general_outbreaks.R
\name{simulate_general_outbreak_inner}
\alias{simulate_general_outbreak_inner}
\title{Generate a single cluster of infections}
\usage{
simulate_general_outbreak_inner(
  cluster_id,
  sample_covariates_df,
  covariate_weights = NULL,
  max_size = 50
)
}
\arguments{
\item{cluster_id}{ID of cluster}

\item{sample_covariates_df}{data frame of covariates with pre-calculated probability of infection \code{prob_inf}}

\item{covariate_weights}{default is NULL which draws uniformly at random with replacement from the sample_covariates_df.  Otherwise, the weights are used.}

\item{max_size}{maximum size of cluster}
}
\value{
data frame with the following columns
\describe{
\item{cluster_id}{unique cluster ID}
\item{person_id}{order of infection in the cluster}
\item{gen}{generation number (>=0)}
\item{inf_id}{ID of the infector}
\item{n_inf}{number of people infected by person}
\item{cluster_pos}{number of positive smears in the cluster}
\item{cluster_size}{number in cluster}
\item{censored}{whether the cluster end was censored or not}
\item{covariates}{other covariates in sample_covariates_df}
}
}
\description{
Generate a single cluster of infections
}
\details{
randomly assigns covariates from sample_df.   breadth not depth.  Generate generation by generation as opposed to going up the branch til termination.
}
