% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_binary_covariate_clusters.R
\name{summarize_binary_clusters}
\alias{summarize_binary_clusters}
\title{Summarize binary covariate clusters}
\usage{
summarize_binary_clusters(df, covariate_name = "x")
}
\arguments{
\item{df}{data frame with the following columns
\describe{
\item{cluster_id}{unique cluster ID}
\item{{covariate_name}}{actual name of feature to summarize over, a binary (0/1) covariate}
}}

\item{covariate_name}{name of the single binary covariate}
}
\value{
a data frame with the following columns
\describe{
\item{freq}{frequency of the following clusters}
\item{cluster_size}{total size of the cluster}
\item{x_pos}{number of individuals in the cluster with the feature of interest =1}
\item{x_neg}{number of individuals in the cluster with the feature of interest = 0}
}
}
\description{
Summarize binary covariate clusters
}
\details{
Condense data from data frames about each individuals to the
summary of the number of indivduals who have a particular covariate feature (1/0).
This assumes the trees are in order by generation.
}
\examples{
example_cluster <- data.frame(cluster_id = c(1, 1, 1,
2, 2,
3, 3, 3, 3,
4),
x = c(0, 1, 1,
0, 0,
1, 0, 1, 1,
0))
summarize_binary_clusters(example_cluster)
}
