% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-mc-binary-cov.R
\name{summarize_binary_cov_trees}
\alias{summarize_binary_cov_trees}
\title{Condense trees into a smaller format}
\usage{
summarize_binary_cov_trees(trees, multiple_outside_transmissions = FALSE)
}
\arguments{
\item{trees}{data frame with the following columns
\describe{
\item{cluster_id}{cluster ID}
\item{n_inf}{number of infected by this individual}
\item{x}{binary covariate value (0/1)}
\item{gen}{generation number}
}}

\item{multiple_outside_transmissions}{default is FALSE}
}
\value{
data frame with the following columns
\describe{
\item{mc_freq}{number of MC trees where this occurred}
\item{cluster_size}{size of cluster}
\item{x_pos}{number of positive x}
\item{x_neg}{number of negative x}
\item{x_pos_trans}{number of transmissions from positives}
\item{x_neg_trans}{number of transmissions from negatives}
\item{root_node}{value of the root node}
}
}
\description{
Condense trees into a smaller format
}
