% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-cov-loglike.R
\name{bp_loglike_binary_cov}
\alias{bp_loglike_binary_cov}
\title{Calculate the likelihood for binary covariate MC trees}
\usage{
bp_loglike_binary_cov(
  inf_params,
  mc_samples_summary,
  return_neg = FALSE,
  multiple_outsider_transmissions = FALSE
)
}
\arguments{
\item{inf_params}{vector of length 2, the beta0 and beta1 coefficients}

\item{mc_samples_summary}{a data frame of sampled transmission trees
with the following columns, where the last one is only needed for the multiple
outsider model.}

\item{return_neg}{logical indicating whether to return the negative loglike.  Default is FALSE}

\item{multiple_outsider_transmissions}{logical indicating whether to use the outsider model likelihood or not}
}
\value{
the estimated log average likelihood over the observed clusters
}
\description{
Calculate the likelihood for binary covariate MC trees
}
