% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-loglike.R
\name{covariate_df_to_mat}
\alias{covariate_df_to_mat}
\title{Take data frame and turn it into matrix for logistic regression}
\usage{
covariate_df_to_mat(mc_trees, cov_names)
}
\arguments{
\item{mc_trees}{data frame with cov_names}

\item{cov_names}{null or string of vector}
}
\value{
matrix of dimension n x (p+1)
}
\description{
Take data frame and turn it into matrix for logistic regression
}
\details{
if cov_names is \code{NA}, then we return a matrix of 1 column of 1s.  This is the case of having \eqn{beta_0} only.  Otherwise, we return the matrix with 1 column of 1s for the intercept and the rest of the variables.
}
