% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_general_cond_trees.R
\name{draw_features}
\alias{draw_features}
\title{Draw the features for the tree people}
\usage{
draw_features(tree_df, feature_type, params_list)
}
\arguments{
\item{tree_df}{data frame where each row is an individual}

\item{feature_type}{currently or supports "single_gauss" "empirical"}

\item{params_list}{additional parameters to pass}
}
\value{
updated data frame with covariate features
}
\description{
Draw the features for the tree people
}
