% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-loglike.R
\name{general_loglike}
\alias{general_loglike}
\title{Calculate the estimated general loglikelihood}
\usage{
general_loglike(
  inf_params,
  observed_data,
  sampled_data,
  return_neg = TRUE,
  cov_mat = NULL,
  cov_names = NULL,
  use_outsider_prob = FALSE,
  return_clust_loglikes = FALSE
)
}
\arguments{
\item{inf_params}{vector of p parameters}

\item{observed_data}{data frame with the following columns
\describe{
\item{cluster_id}{unique cluster ID}
\item{person_id}{order of infection in the cluster}
\item{gen}{generation number (>=0)}
\item{inf_id}{ID of the infector}
\item{n_inf}{number of people infected by person}
\item{censored}{whether the cluster end was censored or not}
\item{cluster_size}{size of the cluster}
\item{covariates}{covariates of the individuals}
}}

\item{sampled_data}{data frame of samples that correspond to the data.  See details.  This is the output of \code{general_cond_tree_sims()}.}

\item{return_neg}{default is TRUE.  Returns the negative loglike}

\item{cov_mat}{optional matrix of covariates corresponding to the sampled_data}

\item{cov_names}{covariate vector of length p which correspond in order to the betas}

\item{use_outsider_prob}{a separate parameter for the outsider infections.  Default is FALSE}

\item{return_clust_loglikes}{if TRUE, function returns the likelihood of every individual cluster}
}
\value{
estimated average loglikelihood for the observed data
}
\description{
Calculate the estimated general loglikelihood
}
\details{
This is a specialized log likelihood function where we first estimate the average log likelihood of trees conditioned by their total size through sampling.  This is very much dependent on the values of sampled_data.
}
