% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_general_cond_trees.R
\name{sample_mc_trees}
\alias{sample_mc_trees}
\title{Sample trees with same features as original but permuted through the trees}
\usage{
sample_mc_trees(observed_data, B = 100, covariate_names = "x")
}
\arguments{
\item{observed_data}{data frame with the following columns
\describe{
\item{cluster_id}{unique cluster ID}
\item{covariates}{covariates of the individuals}
}}

\item{B}{number of samples for each individual tree}

\item{covariate_names}{string names of covariates}
}
\description{
Sample trees with same features as original but permuted through the trees
}
